forgot password done

+++ src/main/java/com/virtu_stock/Security/Authentication/AuthController.java
@@ -88,9 +88,14 @@ public class AuthController {
             @Valid @RequestBody User user, BindingResult result) {
 
         try {
+            if (result.hasErrors()) {
+                Map<String, String> errors = new HashMap<>();
+                result.getFieldErrors().forEach(error -> errors.put(error.getField(), error.getDefaultMessage()));
+                return ResponseEntity.badRequest().body(Map.of("message", errors));
+            }
             if (token == null || token.isBlank()) {
                 return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                        .body(Map.of("message", "Verification token missing"));
+                        .body(Map.of("message", "Invalid Request. Please verify your email first"));
             }
 
             String emailFromToken = jwtUtil.validateOTPToken(token, OTPPurpose.SIGN_UP);
@@ -98,12 +103,7 @@ public class AuthController {
             if (emailFromToken == null || !emailFromToken.equals(user.getEmail())) {
 
                 return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
-                        .body(Map.of("message", "Invalid or expired verification token"));
-            }
-            if (result.hasErrors()) {
-                Map<String, String> errors = new HashMap<>();
-                result.getFieldErrors().forEach(error -> errors.put(error.getField(), error.getDefaultMessage()));
-                return ResponseEntity.badRequest().body(Map.of("message", errors));
+                        .body(Map.of("message", "Invalid email."));
             }
 
             if (!otpService.isEmailVerified(user.getEmail(), OTPPurpose.SIGN_UP)) {
@@ -158,20 +158,22 @@ public class AuthController {
 
         // if (!emailVerificationService.verifyEmail(email)) {
         // return ResponseEntity.badRequest()
-        // .body(Map.of("error", "Invalid email. Please enter a valid email address."));
+        // .body(Map.of("message", "Invalid email. Please enter a valid email
+        // address."));
         // }
         try {
             otpService.generateAndSendOtp(email, OTPPurpose.SIGN_UP);
             return ResponseEntity.ok(Map.of("message", "OTP sent successfully to " + email + ". Valid for 5 minutes"));
         } catch (ResponseStatusException ex) {
             return ResponseEntity.status(ex.getStatusCode())
-                    .body(Map.of("error", ex.getReason()));
+                    .body(Map.of("message", ex.getReason()));
         } catch (MailException ex) {
             return ResponseEntity.status(HttpStatus.BAD_GATEWAY)
-                    .body(Map.of("error", "Failed to send OTP to the provided email. Please check the email address."));
+                    .body(Map.of("message",
+                            "Failed to send OTP to the provided email. Please check the email address."));
         } catch (Exception ex) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(Map.of("error", "An unexpected error occurred while sending OTP."));
+                    .body(Map.of("message", "An unexpected error occurred while sending OTP."));
         }
     }
 
@@ -200,47 +202,64 @@ public class AuthController {
         String email = request.get("email");
         if (email == null || email.isBlank()) {
             return ResponseEntity.badRequest()
-                    .body(Map.of("error", "Email is required"));
+                    .body(Map.of("message", "Email is required"));
         }
         if (!userService.existsByEmail(email)) {
             return ResponseEntity.badRequest()
-                    .body(Map.of("error", "No user found"));
+                    .body(Map.of("message", "No user found"));
         }
         try {
             otpService.generateAndSendOtp(email, OTPPurpose.FORGOT_PASSWORD);
-            return ResponseEntity.ok(Map.of("message", "OTP sent successfully to " + email));
+            return ResponseEntity.ok(Map.of("message", "OTP sent successfully to " + email + ". Valid for 5 minutes."));
         } catch (ResponseStatusException ex) {
             return ResponseEntity.status(ex.getStatusCode())
-                    .body(Map.of("error", ex.getReason()));
+                    .body(Map.of("message", ex.getReason()));
         } catch (MailException ex) {
             return ResponseEntity.status(HttpStatus.BAD_GATEWAY)
-                    .body(Map.of("error", "Failed to send OTP to the provided email. Please check the email address."));
+                    .body(Map.of("message",
+                            "Failed to send OTP to the provided email. Please check the email address."));
         } catch (Exception ex) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(Map.of("error", "An unexpected error occurred while sending OTP."));
+                    .body(Map.of("message", "An unexpected error occurred while sending OTP."));
         }
     }
 
     @PostMapping("/verify-forgot-otp")
     public ResponseEntity<?> verifyForgotOtp(@RequestBody Map<String, String> request) {
-        String email = request.get("email");
-        String otp = request.get("otp");
+        try {
+            String email = request.get("email");
+            String otp = request.get("otp");
 
-        if (email == null || otp == null)
-            return ResponseEntity.badRequest().body(Map.of("error", "Email and OTP are required"));
+            if (email == null || email.isBlank() || otp == null || otp.isBlank()) {
+                return ResponseEntity.badRequest()
+                        .body(Map.of("message", "Email and OTP are required"));
+            }
 
-        boolean verified = otpService.verifyOtp(email, otp, OTPPurpose.FORGOT_PASSWORD);
+            boolean verified = otpService.verifyOtp(email, otp, OTPPurpose.FORGOT_PASSWORD);
 
-        if (verified)
-            return ResponseEntity.ok(Map.of("message", "OTP verified successfully"));
-        else
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                    .body(Map.of("error", "Invalid Email or expired OTP"));
+            if (!verified) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                        .body(Map.of("message", "Invalid or expired OTP"));
+            }
+
+            String token = jwtUtil.generateOTPToken(email, OTPPurpose.FORGOT_PASSWORD);
+
+            return ResponseEntity.ok(
+                    Map.of(
+                            "message", "OTP verified successfully",
+                            "otpToken", token));
+        } catch (ResponseStatusException ex) {
+            return ResponseEntity.status(ex.getStatusCode())
+                    .body(Map.of("message", ex.getReason()));
+        } catch (Exception ex) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body(Map.of("message", "An unexpected error occurred while verifying OTP"));
+        }
     }
 
     @PostMapping("/reset-password")
-    public ResponseEntity<?> resetPassword(@Valid @RequestBody ResetPasswordRequestDTO request, BindingResult result) {
-
+    public ResponseEntity<?> resetPassword(@RequestHeader(value = "x-otp-verify-token", required = false) String token,
+            @Valid @RequestBody ResetPasswordRequestDTO request, BindingResult result) {
         if (result.hasErrors()) {
 
             Map<String, String> errors = new HashMap<>();
@@ -248,30 +267,37 @@ public class AuthController {
 
             return ResponseEntity.badRequest().body(Map.of("errors", errors));
         }
+        if (token == null || token.isBlank()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .body(Map.of("message", "Invalid Request. Please verify your email first"));
+        }
 
         String email = request.getEmail();
         String password = request.getPassword();
-        if (!otpService.isEmailVerified(email, OTPPurpose.FORGOT_PASSWORD)) {
 
-            return ResponseEntity.badRequest().body(
-                    Map.of("error", "Please verify your email before reset password."));
+        String emailFromToken = jwtUtil.validateOTPToken(token, OTPPurpose.FORGOT_PASSWORD);
+
+        if (emailFromToken == null || !emailFromToken.equals(email)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                    .body(Map.of("message", "Invalid email"));
         }
-
+        if (!otpService.isEmailVerified(email, OTPPurpose.FORGOT_PASSWORD)) {
+            return ResponseEntity.badRequest().body(
+                    Map.of("message", "Please verify your email before reset password."));
+        }
         try {
-
             User savedUser = userService.setPassword(email, password);
-
             otpService.deleteByEmailAndPurpose(savedUser.getEmail(), OTPPurpose.FORGOT_PASSWORD);
             mailService.sendPasswordResetMail(savedUser.getEmail(), savedUser.getFirstName());
             return ResponseEntity.ok(Map.of("message", "Password Succesfully updated"));
 
         } catch (UsernameNotFoundException ex) {
             return ResponseEntity.badRequest()
-                    .body(Map.of("error", "User not found"));
+                    .body(Map.of("message", "User not found"));
 
         } catch (Exception ex) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                    .body(Map.of("error", "An unexpected error occurred", "message", ex.getMessage()));
+                    .body(Map.of("status", "An unexpected error occurred", "message", ex.getMessage()));
         }
     }
